+++ packages/_private/angular-demo/src/demofieldset/demofieldset.component.ts
@@ -2,6 +2,7 @@ import { Component, Input } from '@angular/core';
 
 @Component({
   selector: 'demo-fieldset',
+  standalone: false,
   styleUrls: ['./demofieldset.styles.css'],
   templateUrl: './demofieldset.component.html',
 })