+++ packages/components/scripts/vendorism.js
@@ -25,7 +25,6 @@ import {
   vendorMenuItem,
   vendorMenuLabel,
   vendorProgressBar,
-  vendorRadioGroup,
   vendorSelect,
   vendorSpinner,
   vendorTabGroup,
@@ -128,7 +127,7 @@ const otherIncludes = [
 
 const libraryPrefix = 'syn';
 const libraryName = 'synergy';
-const shoelaceVersion = '2.17.1';
+const shoelaceVersion = '2.18.0';
 
 // Command line options
 const optionDefinitions = [
@@ -308,7 +307,6 @@ const config = {
       vendorMenuItem,
       vendorMenuLabel,
       vendorProgressBar,
-      vendorRadioGroup,
       vendorSelect,
       vendorSpinner,
       vendorTabGroup,