+++ packages/components/scripts/vendorism/custom/index.js
@@ -16,7 +16,6 @@ export * from './menu-item.vendorism.js';
 export * from './menu-label.vendorism.js';
 export * from './menu.vendorism.js';
 export * from './progress-bar.vendorism.js';
-export * from './radio-group.vendorism.js';
 export * from './select.vendorism.js';
 export * from './spinner.vendorism.js';
 export * from './tab-group.vendorism.js';